FIX shell plugins installation broken with existing local m2 repo (fixes #45)
Don't be too smart, don't use local m2 repo. We might be smarter and try to use local m2 repo for the artifacts at least to speed up installation on boxes where m2 local repo already has the interesting bits available, but we need a fix version asap, and this is the most simple and stable way to solve the problem.